Make krb5_data compatible with krb5.h
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@1134 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ lib/asn1/der.h
@@ -28,8 +28,8 @@ enum {
 /**/
 
 struct krb5_data {
-  unsigned len;
-  unsigned char *data;
+    size_t len;
+    unsigned char *data;
 };
 
 typedef struct krb5_data krb5_data;
